Moviendo a carpeta y subiendo archivos para trabajar con el mapa
