+++ docs/examples-setup.md
@@ -1,0 +1,43 @@
+# Examples Setup
+
+## iOS
+
+1. If you don't have bundler gem installed:
+
+```
+gem install bundler
+```
+
+2. Install dependencies and open the workspace:
+
+```
+npm install
+npm run build:ios
+npm run run:ios
+```
+
+or
+
+```
+yarn install
+yarn build:ios
+yarn run:ios
+```
+
+## android
+
+1. Start your emulator
+
+2. Install dependencies and run the example:
+
+```
+npm install
+npm run run:android
+```
+
+or
+
+```
+yarn install
+yarn run:android
+```
