--- src/components/DislikeButton.jsx
@@ -1,4 +1,0 @@
-import { customActivityButton } from './ActivityButton';
-
-/** Button to view and perform a "Dislike" action on an item. */
-export default customActivityButton('Dislike', 'Dislike', 'You disliked', 'Disliked');
