--- src/components/FollowButton.jsx
@@ -1,4 +1,0 @@
-import { customActivityButton } from './ActivityButton';
-
-/** Button to view and perform a "Follow" action on an item. */
-export default customActivityButton('Follow', 'Follow', 'You follow', 'Following');
