+++ src/components/ReactComponents/LoggedOut.jsx
@@ -1,0 +1,7 @@
+import useLoggedOut from '../hooks/useLoggedOut';
+
+/** Pane that only shows its contents when the user is logged out. */
+export default function LoggedOut({ children = null }) {
+  const loggedOut = useLoggedOut();
+  return loggedOut ? children : null;
+}
